move Tick to top level
