+++ src/test/scala/polling/BitstampPollingActorSpec.scala
@@ -9,7 +9,8 @@ import akka.stream.scaladsl.{Keep, Source}
 import akka.stream.testkit.scaladsl.{TestSink, TestSource}
 import akka.testkit.TestActorRef
 import akka.util.ByteString
-import co.coinsmith.kafka.cryptocoin.polling.{BitstampPollingActor, BitstampPollingTick, Tick}
+import co.coinsmith.kafka.cryptocoin.Tick
+import co.coinsmith.kafka.cryptocoin.polling.{BitstampPollingActor, BitstampPollingTick}
 import org.apache.avro.generic.GenericRecord
 import org.json4s.JsonAST.JNothing
 import org.json4s.JsonDSL.WithBigDecimal._
